KVM: selftests: arm64: Report set_id_reg reads of test registers as tests
Currently when we run guest code to validate that the values we wrote to
the registers are seen by the guest we assert that these values match using
a KVM selftests level assert, resulting in unclear diagnostics if the test
fails. Replace this assert with reporting a kselftest test per register.

In order to support getting the names of the registers we repaint the array
of ID_ registers to store the names and open code the rest.

Signed-off-by: Mark Brown <broonie@kernel.org>
Message-Id: <20251114-kvm-arm64-set-id-regs-aarch64-v2-1-672f214f41bf@kernel.org>