arm64: RMI: Activate realm on first VCPU run
When a VCPU migrates to another physical CPU check if this is the first
time the guest has run, and if so activate the realm.

Before the realm can be activated it must first be created, this is a
stub in this patch and will be filled in by a later patch.

Signed-off-by: Steven Price <steven.price@arm.com>
Message-Id: <20251217101125.91098-12-steven.price@arm.com>